requirements.txt
asgiref==3.8.1
crispy-bootstrap4==2024.1
crispy-bootstrap5==2024.2
Django==4.2.11
django-crispy-forms==2.1
djangorestframework==3.15.1
pillow==10.3.0
sqlparse==0.5.0
typing_extensions==4.11.0
